# docs/api/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "P0 admin backend for Optimus — auth, RBAC, users, roles, permissions, menus, audit.\nAll authenticated endpoints expect `Authorization: Bearer \u003caccess_token\u003e`.",
        "title": "Optimus Admin API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/credentials/cloud-keys": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "credentials"
                ],
                "summary": "List cloud keys",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "page (default 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "page size (default 20)",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "search by name or description",
                        "name": "q",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "filter by provider (aws|gcp|azure)",
                        "name": "provider",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/optimus-be_internal_infra_response.Envelope"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "credentials"
                ],
                "summary": "Create cloud key",
                "parameters": [
                    {
                        "description": "credential payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_modules_credentials_cloudkey.CreateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/optimus-be_internal_infra_response.Envelope"
                        }
                    }
                }
            }
        },
        "/credentials/cloud-keys/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "credentials"
                ],
                "summary": "Get cloud key",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "credential ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/optimus-be_internal_infra_response.Envelope"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "credentials"
                ],
                "summary": "Update cloud key",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "credential ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "credential payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_modules_credentials_cloudkey.UpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/optimus-be_internal_infra_response.Envelope"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "credentials"
                ],
                "summary": "Delete cloud key",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "credential ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/optimus-be_internal_infra_response.Envelope"
                        }
                    }
                }
            }
        },
        "/credentials/kubeconfigs": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "credentials"
                ],
                "summary": "List kubeconfigs",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "page (default 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "page size (default 20)",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "search by name or description",
                        "name": "q",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "filter by default namespace (exact)",
                        "name": "default_namespace",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/optimus-be_internal_infra_response.Envelope"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "credentials"
                ],
                "summary": "Create kubeconfig",
                "parameters": [
                    {
                        "description": "credential payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_modules_credentials_kubeconfig.CreateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/optimus-be_internal_infra_response.Envelope"
                        }
                    }
                }
            }
        },
        "/credentials/kubeconfigs/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "credentials"
                ],
                "summary": "Get kubeconfig",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "credential ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/optimus-be_internal_infra_response.Envelope"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "credentials"
                ],
                "summary": "Update kubeconfig",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "credential ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "credential payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_modules_credentials_kubeconfig.UpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/optimus-be_internal_infra_response.Envelope"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "credentials"
                ],
                "summary": "Delete kubeconfig",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "credential ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/optimus-be_internal_infra_response.Envelope"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "internal_modules_credentials_cloudkey.CreateRequest": {
            "type": "object",
            "required": [
                "access_key_id",
                "name",
                "provider",
                "secret_access_key"
            ],
            "properties": {
                "access_key_id": {
                    "type": "string",
                    "maxLength": 256
                },
                "description": {
                    "type": "string",
                    "maxLength": 4096
                },
                "name": {
                    "type": "string",
                    "maxLength": 128
                },
                "provider": {
                    "type": "string",
                    "enum": [
                        "aws",
                        "gcp",
                        "azure"
                    ]
                },
                "region": {
                    "type": "string",
                    "maxLength": 32
                },
                "secret_access_key": {
                    "type": "string"
                }
            }
        },
        "internal_modules_credentials_cloudkey.UpdateRequest": {
            "type": "object",
            "properties": {
                "access_key_id": {
                    "type": "string",
                    "maxLength": 256
                },
                "description": {
                    "type": "string",
                    "maxLength": 4096
                },
                "name": {
                    "type": "string",
                    "maxLength": 128
                },
                "provider": {
                    "type": "string",
                    "enum": [
                        "aws",
                        "gcp",
                        "azure"
                    ]
                },
                "region": {
                    "type": "string",
                    "maxLength": 32
                },
                "secret_access_key": {
                    "type": "string"
                }
            }
        },
        "internal_modules_credentials_kubeconfig.CreateRequest": {
            "type": "object",
            "required": [
                "kubeconfig",
                "name"
            ],
            "properties": {
                "default_namespace": {
                    "type": "string",
                    "maxLength": 64
                },
                "description": {
                    "type": "string",
                    "maxLength": 4096
                },
                "kubeconfig": {
                    "type": "string"
                },
                "name": {
                    "type": "string",
                    "maxLength": 128
                }
            }
        },
        "internal_modules_credentials_kubeconfig.UpdateRequest": {
            "type": "object",
            "properties": {
                "default_namespace": {
                    "type": "string",
                    "maxLength": 64
                },
                "description": {
                    "type": "string",
                    "maxLength": 4096
                },
                "kubeconfig": {
                    "type": "string"
                },
                "name": {
                    "type": "string",
                    "maxLength": 128
                }
            }
        },
...
        "optimus-be_internal_infra_response.Envelope": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                },
                "message_key": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and the JWT access token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}